Move Country to models
